Made auth read the body and added some 'optimizations'

+++ notes/main.go
@@ -12,6 +12,7 @@ import (
 
 	"github.com/go-chi/chi/v5"
 	"github.com/go-chi/chi/v5/middleware"
+	"github.com/go-chi/jwtauth/v5"
 	"github.com/joho/godotenv"
 	"go.mongodb.org/mongo-driver/bson"
 	"go.mongodb.org/mongo-driver/bson/primitive"
@@ -34,6 +35,18 @@ type NoteReturn struct {
 	Date        time.Time          `bson:"date"`
 }
 
+var tokenAuth *jwtauth.JWTAuth
+
+func init() {
+	tokenAuth = jwtauth.New("HS256", []byte("authentic"), nil)
+}
+
+/*
+defaults sets default values for the Note struct.
+
+It checks if the Date field is empty and assigns the current time if it is.
+It also checks if the Description field is empty and assigns "N/A" if it is.
+*/
 func (nte *Note) defaults() {
 	var t time.Time
 	if nte.Date == t {
@@ -44,9 +57,26 @@ func (nte *Note) defaults() {
 	}
 }
 
+// main is the entry point of the Go program.
+//
+// main does the following:
+// - Creates a new chi router.
+// - Uses a logger middleware.
+// - Registers a POST route "/new" with the "Create" handler.
+// - Registers a route group "/find" with nested routes.
+//   - Registers a GET route "/{user}/{title}" with the "Read" handler.
+//   - Registers a GET route "/many/{user}/{title}" with the "ReadMany" handler.
+//   - Registers a GET route "/{user}" with the "ReadAll" handler.
+//
+// - Registers a PUT route "/update/{user}/{title}" with the "Update" handler.
+// - Registers a DELETE route "/delete/{user}/{title}" with the "Delete" handler.
+// - Starts the HTTP server and listens on port 3000.
 func main() {
 	r := chi.NewRouter()
 	r.Use(middleware.Logger)
+	r.Use(jwtauth.Verifier(tokenAuth))
+	r.Use(jwtauth.Authenticator(tokenAuth))
+
 	r.Post("/new", Create)
 
 	r.Route("/find", func(r chi.Router) {
@@ -59,6 +89,10 @@ func main() {
 	http.ListenAndServe(":3000", r)
 }
 
+// Create creates a new note in the database.
+//
+// It takes in a http.ResponseWriter and a http.Request as parameters.
+// There is no return value.
 func Create(w http.ResponseWriter, r *http.Request) {
 	var n Note
 	ctx := context.TODO()
@@ -68,6 +102,7 @@ func Create(w http.ResponseWriter, r *http.Request) {
 	}
 	uri, DB, Collection := os.Getenv("URI"), os.Getenv("DB"), os.Getenv("Note")
 	serverAPIOptions := options.ServerAPI(options.ServerAPIVersion1)
+	_, claims, _ := jwtauth.FromContext(r.Context())
 
 	x, err := io.ReadAll(r.Body)
 	if err != nil {
@@ -78,21 +113,33 @@ func Create(w http.ResponseWriter, r *http.Request) {
 	if err != nil {
 		log.Fatal(err)
 	}
-	clientOptions := options.Client().ApplyURI(uri).SetServerAPIOptions(serverAPIOptions)
+	if n.User == claims["username"] {
+		clientOptions := options.Client().ApplyURI(uri).SetServerAPIOptions(serverAPIOptions)
 
-	client, err := mongo.Connect(ctx, clientOptions)
-	if err != nil {
-		log.Fatal(err)
+		client, err := mongo.Connect(ctx, clientOptions)
+		if err != nil {
+			log.Fatal(err)
+		}
+		defer client.Disconnect(ctx)
+		collection := client.Database(DB).Collection(Collection)
+		_, err = collection.InsertOne(ctx, &n)
+		if err != nil {
+			log.Fatal("Could not insert doc")
+		}
+		w.Write([]byte(fmt.Sprintf("Successfully created '%s'", n.Title)))
+	} else {
+		w.Write([]byte("You're not authorized to make this request"))
 	}
-	defer client.Disconnect(ctx)
-	collection := client.Database(DB).Collection(Collection)
-	_, err = collection.InsertOne(ctx, &n)
-	if err != nil {
-		log.Fatal("Could not insert doc")
-	}
-	w.Write([]byte(fmt.Sprintf("Successfully created '%s'", n.Title)))
+
 }
 
+// Read handles the HTTP request and returns a note.
+//
+// The function takes in two parameters:
+//   - w: an http.ResponseWriter object used to write the HTTP response
+//   - r: an *http.Request object representing the HTTP request
+//
+// The function does not return any values.
 func Read(w http.ResponseWriter, r *http.Request) {
 	var n NoteReturn
 	ctx := context.TODO()
@@ -104,6 +151,7 @@ func Read(w http.ResponseWriter, r *http.Request) {
 	}
 	uri, DB, Collection := os.Getenv("URI"), os.Getenv("DB"), os.Getenv("Note")
 	serverAPIOptions := options.ServerAPI(options.ServerAPIVersion1)
+	_, claims, _ := jwtauth.FromContext(r.Context())
 
 	clientOptions := options.Client().ApplyURI(uri).SetServerAPIOptions(serverAPIOptions)
 
@@ -112,22 +160,34 @@ func Read(w http.ResponseWriter, r *http.Request) {
 		log.Fatal(err)
 	}
 	defer client.Disconnect(ctx)
-	filter := bson.D{{Key: "title", Value: t}, {Key: "user", Value: u}}
-	collection := client.Database(DB).Collection(Collection)
-	c, err := collection.Find(ctx, filter)
-	if err != nil {
-		log.Fatal("Could not find any matches")
-	}
-	defer c.Close(context.TODO())
-	for c.Next(context.TODO()) {
-		err := c.Decode(&n)
+
+	if u == claims["username"] {
+		filter := bson.D{{Key: "title", Value: t}, {Key: "user", Value: u}}
+		collection := client.Database(DB).Collection(Collection)
+		c, err := collection.Find(ctx, filter)
 		if err != nil {
-			log.Fatal("Could not decode a match")
+			log.Fatal("Could not find any matches")
 		}
+		defer c.Close(context.TODO())
+		for c.Next(context.TODO()) {
+			err := c.Decode(&n)
+			if err != nil {
+				log.Fatal("Could not decode a match")
+			}
+		}
+		json.NewEncoder(w).Encode(&n)
+	} else {
+		w.Write([]byte("You're not authorized to make this request"))
 	}
-	json.NewEncoder(w).Encode(&n)
 }
 
+// ReadMany retrieves multiple documents from the database based on the provided title and user.
+//
+// It takes in the following parameters:
+// - w: http.ResponseWriter - The response writer used to write the response back to the client.
+// - r: *http.Request - The HTTP request received from the client.
+//
+// The function does not return any value.
 func ReadMany(w http.ResponseWriter, r *http.Request) {
 	var n []bson.M
 	ctx := context.TODO()
@@ -139,6 +199,7 @@ func ReadMany(w http.ResponseWriter, r *http.Request) {
 	}
 	uri, DB, Collection := os.Getenv("URI"), os.Getenv("DB"), os.Getenv("Note")
 	serverAPIOptions := options.ServerAPI(options.ServerAPIVersion1)
+	_, claims, _ := jwtauth.FromContext(r.Context())
 
 	clientOptions := options.Client().ApplyURI(uri).SetServerAPIOptions(serverAPIOptions)
 
@@ -150,19 +211,26 @@ func ReadMany(w http.ResponseWriter, r *http.Request) {
 
 	collection := client.Database(DB).Collection(Collection)
 
-	filter := bson.D{{Key: "title", Value: bson.D{{"$regex", "^" + t}}}, {Key: "user", Value: u}}
-	fmt.Print(filter)
-	c, err := collection.Find(ctx, filter)
-	if err != nil {
-		log.Fatal("Could not find any matches")
+	if u == claims["username"] {
+		filter := bson.D{{Key: "title", Value: bson.D{{"$regex", "^" + t}}}, {Key: "user", Value: u}}
+		c, err := collection.Find(ctx, filter)
+		if err != nil {
+			log.Fatal("Could not find any matches")
+		}
+		defer c.Close(context.TODO())
+		if err = c.All(ctx, &n); err != nil {
+			log.Fatal(err)
+		}
+		json.NewEncoder(w).Encode(&n)
+	} else {
+		w.Write([]byte("You're not authorized to make this request"))
 	}
-	defer c.Close(context.TODO())
-	if err = c.All(ctx, &n); err != nil {
-		log.Fatal(err)
-	}
-	json.NewEncoder(w).Encode(&n)
 }
 
+// ReadAll handles the HTTP request to read all documents for a specific user.
+//
+// It takes in the http.ResponseWriter and http.Request as parameters.
+// It does not return anything.
 func ReadAll(w http.ResponseWriter, r *http.Request) {
 	var n []bson.M
 	ctx := context.TODO()
@@ -173,6 +241,7 @@ func ReadAll(w http.ResponseWriter, r *http.Request) {
 	}
 	uri, DB, Collection := os.Getenv("URI"), os.Getenv("DB"), os.Getenv("Note")
 	serverAPIOptions := options.ServerAPI(options.ServerAPIVersion1)
+	_, claims, _ := jwtauth.FromContext(r.Context())
 
 	clientOptions := options.Client().ApplyURI(uri).SetServerAPIOptions(serverAPIOptions)
 
@@ -184,19 +253,30 @@ func ReadAll(w http.ResponseWriter, r *http.Request) {
 
 	collection := client.Database(DB).Collection(Collection)
 
-	filter := bson.D{{Key: "user", Value: u}}
-	fmt.Print(filter)
-	c, err := collection.Find(ctx, filter)
-	if err != nil {
-		log.Fatal("Could not find any matches")
+	if u == claims["username"] {
+		filter := bson.D{{Key: "user", Value: u}}
+		c, err := collection.Find(ctx, filter)
+		if err != nil {
+			log.Fatal("Could not find any matches")
+		}
+		defer c.Close(context.TODO())
+		if err = c.All(ctx, &n); err != nil {
+			log.Fatal(err)
+		}
+		json.NewEncoder(w).Encode(&n)
+	} else {
+		w.Write([]byte("You're not authorized to make this request"))
 	}
-	defer c.Close(context.TODO())
-	if err = c.All(ctx, &n); err != nil {
-		log.Fatal(err)
-	}
-	json.NewEncoder(w).Encode(&n)
 }
 
+// Update updates a note in the database based on the provided title and user.
+//
+// Parameters:
+// - w: the http.ResponseWriter object for writing the response.
+// - r: the *http.Request object representing the incoming request.
+//
+// Return type:
+// This function does not return anything.
 func Update(w http.ResponseWriter, r *http.Request) {
 	var n Note
 	ctx := context.TODO()
@@ -208,6 +288,7 @@ func Update(w http.ResponseWriter, r *http.Request) {
 	}
 	uri, DB, Collection := os.Getenv("URI"), os.Getenv("DB"), os.Getenv("Note")
 	serverAPIOptions := options.ServerAPI(options.ServerAPIVersion1)
+	_, claims, _ := jwtauth.FromContext(r.Context())
 
 	x, err := io.ReadAll(r.Body)
 	if err != nil {
@@ -229,15 +310,27 @@ func Update(w http.ResponseWriter, r *http.Request) {
 
 	collection := client.Database(DB).Collection(Collection)
 
-	filter := bson.D{{Key: "title", Value: t}, {Key: "user", Value: u}}
-	update := bson.D{{"$set", bson.D{{Key: "title", Value: n.Title}, {Key: "description", Value: n.Description}, {Key: "date", Value: n.Date}}}}
-	_, err = collection.UpdateOne(ctx, filter, update)
-	if err != nil {
-		log.Fatal(err)
+	if u == claims["username"] {
+		filter := bson.D{{Key: "title", Value: t}, {Key: "user", Value: u}}
+		update := bson.D{{"$set", bson.D{{Key: "title", Value: n.Title}, {Key: "description", Value: n.Description}, {Key: "date", Value: n.Date}}}}
+		_, err = collection.UpdateOne(ctx, filter, update)
+		if err != nil {
+			log.Fatal(err)
+		}
+		w.Write([]byte(fmt.Sprintf("Successfully updated '%s'", n.Title)))
+	} else {
+		w.Write([]byte("You're not authorized to make this request"))
 	}
-	w.Write([]byte(fmt.Sprintf("Successfully updated '%s'", n.Title)))
 }
 
+// Delete deletes a document from the specified collection in the MongoDB database.
+//
+// Parameters:
+// - w: The http.ResponseWriter used to write the response back to the client.
+// - r: The *http.Request containing the HTTP request details.
+//
+// Returns:
+// The function does not return anything.
 func Delete(w http.ResponseWriter, r *http.Request) {
 	ctx := context.TODO()
 	t := chi.URLParam(r, "title")
@@ -251,6 +344,7 @@ func Delete(w http.ResponseWriter, r *http.Request) {
 	uri, DB, Collection := os.Getenv("URI"), os.Getenv("DB"), os.Getenv("Note")
 	serverAPIOptions := options.ServerAPI(options.ServerAPIVersion1)
 	clientOptions := options.Client().ApplyURI(uri).SetServerAPIOptions(serverAPIOptions)
+	_, claims, _ := jwtauth.FromContext(r.Context())
 
 	client, err := mongo.Connect(ctx, clientOptions)
 	if err != nil {
@@ -260,9 +354,13 @@ func Delete(w http.ResponseWriter, r *http.Request) {
 
 	collection := client.Database(DB).Collection(Collection)
 
-	_, err = collection.DeleteOne(ctx, filter)
-	if err != nil {
-		log.Fatal(err)
+	if u == claims["username"] {
+		_, err = collection.DeleteOne(ctx, filter)
+		if err != nil {
+			log.Fatal(err)
+		}
+		w.Write([]byte(fmt.Sprintf("Successfully deleted '%s'", t)))
+	} else {
+		w.Write([]byte("You're not authorized to make this request"))
 	}
-	w.Write([]byte(fmt.Sprintf("Successfully deleted '%s'", t)))
 }
